requirements.txt
GitPython==3.1.32
beautifulsoup4==4.12.2
scrapy==2.9.0
spacy==3.6.1
transformers==4.33.2
openai==1.3.9
apscheduler==3.10.4
Markdown==3.5.1
reportlab==4.0.7
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1 